unattended-file: Add a forgotten 'else'
In commit ae2bbc81, I was supposed to add an 'else if' but I ended-up
creating a new 'if' block by forgetting to type 'else'. This made Boxes
always choose floppy injection method if script claimed to support it
and that broke express installation for scripts that pretend to support
floppy method but actually don't (Fedora and RHEL).