machine: Ensure 'status' property is always set
We were not initializing the 'status' property and not resetting it
later either so it'll remain to be 'Connecting to MACHINE_NAME' even
after connection.

This isn't exactly an issue on its own but later we'll indirectly
connect this property to the a11y-name of the containing window and that
property is not nullable.