topbar: Ensure 'status' property is always set
We were not initializing the 'status' property and resetting it
to null in collection view. This meant an empty titlebar in collection
view. Apart from this not being a good thing (see bug below), later we'll
connect this property to the a11y-name of the containing window and that
property is not nullable.

https://bugzilla.gnome.org/show_bug.cgi?id=736972