resource-graph: Adaptive ymax
For the case of data being I/O bytes written/read, we can't just assume a
particular fixed value for max on y-axis (ymax) since I/O speed varies a
lot across different hardware. The result of a fixed ymax is typically
graphs fluctuating between 0 and ymax and that is not just useless but
also looks ugly.

Lets reset ymax to the value of data point if it happens to be greater
than current ymax.