vm-creator: Set autostart=true on storage pool
Even though we ensure that storage pool in activated if its been
de-activated for some reason and even create the backing directory, it
seems that on at least some machines we fail to do that somehow.

With lack of any data on why and when that happens, let's make the
storage pool autostart in hopes it helps the situation.

https://bugzilla.gnome.org/show_bug.cgi?id=748160