app-window: Add 'collection_view' property
Rename the 'view' attribute to 'collection_view', making it private in
the process, and add a new generic 'view' property that always points to
the current collection view of the window.

For now this isn't very useful since we only have a single view but a
list view will be added in a following patch.

https://bugzilla.gnome.org/show_bug.cgi?id=733252