i-props-provider: Size props' name & value together
We center the size props' value label in hopes that it'll align nicely
with I/O graph's label and with eachother but we end up with misaligned
size value labels, which looks really bad and quite contrary to UI
mockup.

I failed to find a way to make proper centering happen so let's just
place the size value labels next to their respective name labels. This
doesn't look as good as the UI mockup but still much better than the
misaligned labels we have.