Revert "machine: Ensure 'status' property is always set"
This reverts commit 63b978a880f8e72d8353a5b5c50223bbf09a489e.

This retioanale for the reverted commit was the fact that 'status'
property is (indirectly) bound 'accessible-name' property of AppWindow
and that is non-nullable so we end-up with a warning on console. This
issue will be solved in a following patch in a different way but we
can't have Machine.name being permanently mapped to it's 'status'
anymore as we are about to expand the role of 'status' property and show
it's contents also in collection view.

https://bugzilla.gnome.org/show_bug.cgi?id=753888