machine: Status in 'status' prop, not 'info'
Since 'info' property is bound to 2nd text line under the box thumbnail,
we have been (ab)using this property to keep status messages too. We
already have a 'status' property for those and that is where these
messages should go into.

With this patch, these status messages will no longer appear under the
box thumbnail but this is a temporary regression that will be fixed in a
following patch.

https://bugzilla.gnome.org/show_bug.cgi?id=753888