vm-configurator: Always update network config
Instead of only updating network config for existing domains if bridge
networking is available and existing networking in the box is not
already bridge network, let's always do so.

This fixes the issue of VM not starting if bridge networking is no
longer available (this has been known to happen due to setup/packaging
issues) and existing network is set to either bridge or some other
non-working configuration (e.g 'default' when importing from system
libvirt).

https://bugzilla.gnome.org/show_bug.cgi?id=756547