graph: Use style's state to draw rather than widget's
Passing a fixed state to Gtk.StyleContext.get_color() will result in
warnings on the console with latest gtk+ and is not supported way of
doing things anymore. This patch sets the required state on the style and
then uses the style's state to get the color for that state.

https://bugzilla.gnome.org/show_bug.cgi?id=763677