wizard-source: Keep URI on cleanup if on URL page
Navigating back to the source page in wizard, has the side effect of
cleaning-up the wizard source. This means also clearing the URI, which
is very unfriendly to user who could have painstakenly hand written the
URI but made a small mistake while doing so, and wants to go back to
simply fix his/her minor mistake.

So let's not clear the UI as part of wizard source clean-up if source is
on URL page.

https://bugzilla.gnome.org/show_bug.cgi?id=747855