app: Simply quit() on removal of main window
Our shutdown() implementation needs to cleanup/shutdown all windows and
it won't do that if we already remove the main window from the list
of all the windows (that shutdown() iterates over) before launching the
shutdown process.

This fixes the issue of notifications not getting dismissed (and hence
deletion of boxes not happning for deletion notifications) if user quits
Boxes by pressing the 'x' button in the titlebar, instead of Ctrl+q
shortcut.

https://bugzilla.gnome.org/show_bug.cgi?id=761479