libvirt-machine-props: Allow resize of disk w/ snapshots
Since it is currently not possible to resize a disk image if it has
snapshots on it, resize operation ends up not working with a warning on
the console. Let's fix that by deleting all associated snapshots before
attempting to resize the disk image but not before notifying user about
what we are about to do.

https://bugzilla.gnome.org/show_bug.cgi?id=756140