wizard: Ignore unimportable VMs early
We currently assume all system libvirt VMs to be importable and end up
providing an option to import unimportable VMs and failing to import
them when user chooses that option. This looks really bad.

Let's check the existence of the VMs' disk images and ensure their
readability, early on and simply ignore the VMs without a valid disk
image attached to them.

https://bugzilla.gnome.org/show_bug.cgi?id=754500