installed-media: Allow bypassing format check
In certain cases (e.g the cloning code that will be added in a following
patch), the media will already be in qcow2 format so there is no need to
check for format of disk image. Moreoever, currently our disk images
lack the file extension so our own code breaks on detecting it's type as
we rely on file extensions for detection.

https://bugzilla.gnome.org/show_bug.cgi?id=710307