system-importer: Ensure disk reability on import
Ensuring disk readability involves elevation of privileges so if we do
that on startup, we end up launching a polkit dialog even if user never
uses the option of libvirt system VMs.

While we should still check for disk images before showing the option to
import their VMs from system libvirt, it's best if we only ensure their
reabibility when user explicitly asks for import.

https://bugzilla.gnome.org/show_bug.cgi?id=766727