Remove smartcard hacks for old Qemu
Just like the previous patch, where we removed hacks for lack of USB
direction in Qemu < 1.3, in this patch we do the same for smartcards. I do
not know when smartcard support was added in Qemu but it was definitely
even before USB redirection so this is even safer than USB redirection
patch.