remote-machine: Connect to a non-null unconnected display
The problem is that, when connecting to a display, we only connect
to it if it hasn't already been created. By doing this, we ignore
the case when a display is created but not connected, which is a
logical scenario and can lead to a grey screen by choosing not to
connect.

This scenario happens already when fetching the display's properties,
now that we keep the display alive if USB devices are available.

To avoid this issue, let's not assume that the device is already
connected just because it has been created.

https://bugzilla.gnome.org/show_bug.cgi?id=744905