app-window: Disconnect from machine callbacks at removal
The problem is that after removing a window, its deleted machine
is still able to send signals to it and consequentially to try
to remove it again during the actual machine deletion process,
which leads to undefined behaviour.

Let's simply disconnect the machine's callbacks in its window
when the machine is signaled as deleted.

https://bugzilla.gnome.org/show_bug.cgi?id=767214