display-page: Don't show widgets from size-allocate signal
Don't show widgets from size-allocate signal handler cause that causes
Gtk+ to show warnings:

(gnome-boxes:1480): Gtk-WARNING **: GtkGrid 0xa73cc0 is drawn without a current allocation. This should not happen.

and also with latest Gtk+ the showing of display size was not working. I
also suspect some of the glitches I saw, to be coming from this check in
Gtk+.