libvirt-machine: Don't autosave live & installation
Express installation is not available for all OSes so it's not a very rare
case to have multiple (non-express) installations running in background.
Also it installation from a live session is a pretty common case now. So
let's just not do autosave for all live and under-installation VMs, not
just express installed ones.