unattended-setup-box: Don't cache passwords in plain text
The user password specified during an express installation is cached
in plain text in ~/.config/gnome-boxes/unattended/setup-data.conf.
Instead, we should use the keyring - it's meant for this. As long as
the host user account has a password, the keyring will be encrypted
and protected in case the disk is stolen.

We store the password as a serialized GVariant dictionary because it
is more extensible than a simple string. In future we might want to
store multiple secrets for a particular media (eg., registration
passwords, product keys, etc.), and those can be accommodated by
specifying a different key in the dictionary.

Fallout from ae21e562515ab60d3279f182313e9a6c7be53abd

https://bugzilla.gnome.org/show_bug.cgi?id=777788