app, window: Introduce troubleshoot view
Sometimes we are not able to launch the application due to issues
down the stack that we cannot ignore in Boxes.

This patch introduces an error message when Boxes is not able to
start. In the future it could launch GNOME Logs. This is intented
for errors we cannot recover from.

https://bugzilla.gnome.org/show_bug.cgi?id=784896