build: Fix the Makefile rules for the GSettings schema
@GSETTINGS_RULES@ comes from the GLIB_GSETTINGS M4 macro shipped by
GLib in gsettings.m4. It's unrelated to intltool and should be retained
as long as Makefiles are being used.

Secondly, unlike desktop and appdata files, the translations for
GSettings schemas are not merged back during the build. The
translatable strings are extracted during the build, but the
translations are fetched at runtime. This is similar to how it works
for source code. Therefore, there is no need for the *.gschema.xml.in
to *.gschema.xml indirection. Note that this was already not necessary
with intltool >= 0.50.

Fallout from 10476bf2cb47e8fcea995a9faf595c685e88d4b5

https://bugzilla.gnome.org/show_bug.cgi?id=783832