wizard-source: Separate the groups of sources using headings
This increases the height of the WizardWindow, so the WizardScrolled
needs to scroll more aggressively.

Note that the vertical height of the window is still more than what it
used to be. That will be addressed in the following patch by moving the
file selection button next to the heading label.

https://bugzilla.gnome.org/show_bug.cgi?id=786679