downloader, wizard: Support a custom filename for downloaded media
The URI for getting gratis RHEL ISOs from the Red Hat developer portal
at developers.redhat.com looks like this:
  https://access.cdn.redhat.com/.../rhel...dvd.iso?_auth_=......

While it seems possible to get the basename of the downloaded file by
removing the URI's query string, there is no guarantee that the
structure of the URI won't change in future.

On the other hand, the basename of the downloaded ISO is encoded in the
authentication URI that's embedded in the initial JSON or XML response
from the server. It might be better to use that instead. Therefore,
there is a need to explicitly specify the filename.

https://bugzilla.gnome.org/show_bug.cgi?id=786679