configure: Require gtk+ >= 3.22.20
In a following patch, we will need to use GtkBuilder's "type-func"
attribute to override its built-in heuristics. Prior to this,
GtkBuilder wouldn't let the "type-func" override the heuristics if the
"class" attribute was also present. On the other hand, older Vala
compilers treat the "class" attribute as mandatory.

It's slightly more convenient to bump the gtk+ requirement than Vala
because gtk+-3.22.x is a long-term stable series, so older
distributions are likely to keep it updated.

See:
https://bugzilla.gnome.org/show_bug.cgi?id=786932
https://bugzilla.gnome.org/show_bug.cgi?id=787033
https://bugs.webkit.org/show_bug.cgi?id=175937

https://bugzilla.gnome.org/show_bug.cgi?id=786679