gtk-style: Apply proper border to boxes-menu-row
The "Enter URL" button in the New Box wizard had a border-radius
applied to its bottom instead of the top.

The intention of this styling is to indicate a correlation between
the button and the content below. This way, the border-radius
should be applied in the top as the mockups indicate.

https://bugzilla.gnome.org/show_bug.cgi?id=787527