wizard: Always use media-optical when downloading, not some bogus logo
If the user temporarily selected any InstallerMedia, either from the
file chooser or the list of local images, before deciding to download
one from a URI, the logo wouldn't be reset. So, if a Fedora URI was
preceded by selecting an Ubuntu image, then it would continue to show
the Ubuntu logo while downloading Fedora.

Since the "prepare" method is invoked twice for URIs requiring a
download - once before initiating the download and once after, it is
better to reset the icon even earlier.

https://bugzilla.gnome.org/show_bug.cgi?id=777133