wizard, wizard-source: Don't align & margin WizardSource
In a following patch, we will add a separate page with a WebKitWebView
to the WizardSource. When used, this page should expand to fill the
whole dialog instead of staying centre aligned like the existing pages.
Therefore, each page in WizardSource needs to be margined and aligned
separately, which is already the case for the horizontal alignment.

This fixes the vertical alignment and the horizontal margins.

Note that this has a subtle effect on the layout of the WizardSource
pages. So far, WizardSource was vertically centred and url_menubox
was set to fill, and WizardSource was vertically taller than the URL
page. This meant that url_menubox would expand vertically to fill the
entire WizardSource, but due to its semantics as a vertically oriented
GtkBox, it's contents would stick to it's top edge. This gave the
GtkLabel and GtkEntry widgets in the page their slightly off-centred
positioning on the vertical axis.

Moving the valign=CENTER from the WizardSource to url_menubox would
spoil the layout because it will force url_menubox to be strictly
centred and non-expanding. Therefore, top and bottom margins have been
used instead of alignment. The existing top margin was increased to
match that of the main page.

To prevent the top edge of the widgets from jumping when changing
pages, the main page now has a similar layout as the URL page. It is
no longer centred and has a similarly off-centred positioning on the
vertical axis. However, this is much more subtle and barely
discernible.

https://bugzilla.gnome.org/show_bug.cgi?id=786679