actions-popover: Add "Send file" option
The problem is that with the current file transferring
DnD mechanism, users may not be aware of the fact that they
are able to send files to guests which have this ability.

Using this option, in following patches users will also be
able to transfer files to a SPICE guest using a more intuitive
click of a button.

https://bugzilla.gnome.org/show_bug.cgi?id=762321