display-page: Show drag and drop message
Currently, when transferring files through DnD, the user
isn't aware of the actual destination of the files and doesn't
receive any visual feedback.

Let's show a message when the user is hovering with a file over a
display which can receive files(currently only SPICE), informing him
that he is about to send files to the machine and pointing to him
the destination of the files.

https://bugzilla.gnome.org/show_bug.cgi?id=762321