build: Use PACKAGE_VERSION instead of VERSION
One part of it is purely cosmetic. The PACKAGE_ prefix increases the
name's uniqueness, hence making it more greppable, and goes nicely
with the other variables with the same prefix.

A following commit will add an option to specify the distributor's
version. Using PACKAGE_VERSION instead of VERSION removes any shadow
of doubt as to whose version it is.

https://bugzilla.gnome.org/show_bug.cgi?id=790343