Set a user-agent when using the network
The user-agent looks like this:
  <distro-name>/<distro-version> <package-tarname>/<package-version>

The user agent is used with the Soup.Session when downloading ISOs and
assets (like logos), and passed as the "tag" parameter to the
developers.redhat.com end-point.

<distro-name> and <distro-version> default to "GNOME" and "3"
respectively. Build options have been added to enable distributors to
suitably override them.

<package-tarname> is always "gnome-boxes", and <package-version> is
determined by git-version-gen. So, tarball releases get a nice round
number based on the name of the tag, while development builds get a
unique version string based on the last commit hash. One drawback is
that distributors can't override the <package-version> for downstream
patches. An option can be added in future if this becomes a problem.

https://bugzilla.gnome.org/show_bug.cgi?id=790343