wizard-source: Compose a human-readable media label
To enumerate the medias of an operating system in a way that it
isn't the literal file name neither a duplicated name, we should
rely on the Osinfo.OsVariant "name" property.

Unfortunately the Osinfo database doesn't contain proper names
for OS variants. This way, these change introduce some heuristic
to handle the most common iso file names for the operating systems
available to download.

If there will be a case not covered by this heuristics, we present
the literal iso file name with %s/_/" "/g.

This commit should be reverted when Osinfo contains distinguishable
variant names.

https://bugzilla.gnome.org/show_bug.cgi?id=790614