meson: Use LINGUAS file for help generation
The linguas parameter is deprecated in gnome's yelp function for
the future meson's versions.

The patch uses the LINGUAS file instead of the linguas parameter.
It also bumps meson's version to support this feature properly.

The polish translations has also been added as it was present in
autotools but not in meson.

https://bugzilla.gnome.org/show_bug.cgi?id=791421