build: Rename Boxes desktop file
The desktop file has two `.in` extensions due to the two steps
to be perfomerd: variable substitution and translation merge.

However, there are no variables to be substituted so one of the
`.in` extensions can be removed.

This patch renames the file leaving only a `.in` extension which
is used as an indication that translations must be merged.

https://bugzilla.gnome.org/show_bug.cgi?id=791421