collection: Decorate a Gtk.ListStore to manage items
Using a Gtk.ListStore we can let Gtk.ListBox and Gtk.FlowBox
bind to the same model, and apply both filtering and sorting
at the same place.

These changes require project-wide tweaks in the way we iterate
over CollectionItems. For this, a Collection.foreach was introduced.
We use the classic C for loop for the other cases where loop control
instructions are necessary.

A step back in this implementation in comparison to the previous
one is that now the removal of items is done in linear time.

https://bugzilla.gnome.org/show_bug.cgi?id=791839