list-view: Bind GtkListBox to Collection model
We used to subscribe to the item_added and item_removed signals of
Collection and insert the rows accordingly.

Now the ListView listbox is binded directly to the Collection model
and therefore the model-view mapping is done automatically.

The changes in app.vala and app-window.vala are temporary so we
don't brake the current implementation of IconView.

https://bugzilla.gnome.org/show_bug.cgi?id=791839