icon-view: Port to GtkFlowBox
Finally IconView can be directly binded to Collection the same way
as ListView is. This is one important step towards providing full
independence between the model and the views.

We can also now revert the app.vala, app-window.vala hack introduced
before to prevent the IconView to break while ListView was being
ported.

https://bugzilla.gnome.org/show_bug.cgi?id=791839