collection: Handle hidden items at the model side
A hidden item is a newly create item (machine) where the user is
still visualizing the wizard, and therefore should'nt be presented
in the views.

Both views (IconView and ListView) used to handle the hidden_items
independently, duplicating the code and logic.

Now we can handle the hidden_items in a single place, in the Collection
object as a result of the decoupling of model and view recently
introduced.

https://bugzilla.gnome.org/show_bug.cgi?id=791839