meson: Use meson's convention
The convention when naming dependencies is to use the '_dep' suffix,
which helps detecting dependency variables. The variables have been
renamed to follow this convention.

The dependencies that are also used once, are declared in the
target's dependency array which helps meson processing less and also
one less variable is needed.

https://bugzilla.gnome.org/show_bug.cgi?id=791421