meson: Use glib's version as target glib
meson's is able to resolve the expected vala `target-glib` by using
glib's version.

At the moment, it appends the 2.38 as the `target-glib` but also
2.44 is also added by using vala's argument.

The glib's and gio's versions have been raised to 2.44 along what is
used by gobject dependency, and the vala argument is removed, so
meson itself is able to add the correct `target-glib` version.

https://bugzilla.gnome.org/show_bug.cgi?id=791421