notificationbar: Call each notification dismiss() at cleanup
Each Notification object wraps in a GLib.Timeout in order to
schedule the execution of the callback.

Boxes gets stuck if it is closed when a notification hasn't
finished yet.

The solution for that is to make sure that each Notification's
dimiss() method gets called when the GtkApplication shutdown
method is invoked. This way we can ensure that all the sources
are removed from the default main context.

https://bugzilla.gnome.org/show_bug.cgi?id=784923