actions-popover: Drop "Pause" option
The "Pause" option refers to the libvirt action of "saving a guest",
which consists of taking a running guest and save its memory state
to a file, enabling it to be later restored causing the guest to
recover its original running state. See [0].

Boxes will, by default, Pause/Save a guest whenever its display is
closed/destroyed, unless when the box is marked to "Run in Background".

This way, there's no need to expose the "Pause" shortcut in the actions
popover.

Besides, we already expose the "Pause" action in the action bar,
whenever the selection mode is active. This can be very convenient for
users that want to Pause multiple guests simultaneously.

[0] https://libvirt.org/guide/html/Application_Development_Guide-Lifecycle-Save.html

Resolves #139