wizard: Center the "Customize" button below
The current placement of the "Customize" button makes it seems
like the action is tied to the Disk property only.

The wireframes for the new machine assistant already position the
button in a row below, expanding horizontally to the size of the
description text in the top.

https://raw.githubusercontent.com/gnome-design-team/
gnome-mockups/master/boxes/wires/newbox-assistant.png

Fixes #71