wizard: Show logo while downloading image
This fixes a regression introduced with the "Download an OS"
feature.

When introducing the option to download RHEL we were populating a
media_urls_store which was later queried to provide a Osinfo.Os
given a download URL.

The media_urls_store was removed when the URL completion was dropped
in the Enter a URL view, since it was made obsolete by the "Download
an OS" page.