wizard: Notify when unable to add RDP server
Whenever Boxes is compiled without RDP support (rdp=false build
option) and the user wants to add an rdp:// address, we present
an human-readable error message notifying that.

This is a counter measure because we want to advertise the RDP
support in the "Enter an URL" dialog, but the RDP support is
optional (build option).