ci: Generate a Flatpak bundle for every commit/merge_request
See https://gitlab.gnome.org/GNOME/gnome-boxes/issues/202

This way a flatpak bundle gets generated and linked in the GitLab
merge requests for developers to test it.

The flatpak bundle is a gnome-boxes.flatpak file that can be
installed with:
  $ flatpak install gnome-boxes.flatpak

And can run alongside your existent Boxes install.

A bundle is available for 30 days until it expires.

We still build every merge request against a Fedora image to
guarantee it builds there.

Fixes #202