wizard: Set placeholder text in Download an OS search
The placeholder text gets removed when the entry receives focus
and we want users to be able to type without having to click/focus
in the search entry. That's why we work it around by handling key
press events at a parent widget level, passing the event along to
the GtkSearchEntry (downloads_search) widget when the Wizard page
is the Downloads one.

This is based on the mockup available at
https://gitlab.gnome.org/Community/Design/app-mockups/raw/87e07fb/boxes/boxes-newbox-assistant.png