wizard-downloads-page: Bump recommended downloads list
We should offer the latest versions of each of the recommended
OSes.

For the future releases it would be better to develop a process
of finding the latest release of each entry automatically. Also
downstreams should have an easier way to patch the list to
include offers of their choice.