flatpak: Bump flatpak runtime version
The flatpak manifest is used to generate the nightly flatpak build,
but currently it uses the 3.28 runtime, rather than the nightly
master. As a result it tries to build on outdated meson, and fails.

This patch updates to the master runtime, which will make sure there
aren't any build failures before the 3.30 release.