vm-configurator: Hardcode xmlns URI
Don't use PACKAGE_URL; if it changes in the future and this usage
site isn't change, boxes won't be able to read it's pre-existing
VM metadata. In fact it appears this already happened, see
912ff3ec520e

An xmlns URI doesn't need to actually be resolvable, just unique.
So let's hardcode the current value and never change it