flatpak: Fix the build on Aarch64
The mtools release is old and predates the Aarch64 support in the
autotools.

As a result, the included config.{guess,sub} files do not know about the
Aarch64 architecture, and the build fails.

This commit takes some newer config.{guess,sub} files which the Sdk
helpfully provides for exactly this purpose.