collection-toolbar: Prevent Gtk+ from creating app-menu
AppMenus are going away in GNOME 3.32. So far Boxes was relying on
Gtk+ to discover the "gtk/menus.ui" resource and create an appmenu
from it.

The deprecation strategy will add existing app-menus automatically
in the headerbar, in order to avoid introducing accessibility
regressions.

This way, we should prevent Boxes' current menu from being mistakenly
picked up by GtkApplication as a legacy appmenu. We can easily
achieve this by renaming the menu resource away from "gtk/menus.ui".

See the "Automatic resources" section in the GtkApplication
documentation for details of this behavior.
https://developer.gnome.org/gtk3/stable/GtkApplication.html

This is a complementary effort of the AppMenu retirement initiative.
https://gitlab.gnome.org/GNOME/Initiatives/wikis/App-Menu-Retirement

Fixes #223