libvirt-machine: Recreate network interface while cloning
While cloning a libvirt machine, Boxes copies the literal VM XML
config and duplicates it into a newly created machine. This way
the machines are expected to be identical.

Users won't ever want to have two Boxes with the same mac address.

A common use-case for Boxes is for someone to create and configure
a single box and clone it into multiple instances.

Therefore after duplicating the machine's XML, we now re-add the
network interface, this way the cloned machine will have a different
MAC address than the original one.

Fixes #262