properties-window: Use GtkFileChooserNative
These changes are necessary in order to make Boxes consume the
Flatpak file chooser portal.

FileChooserNative is an abstraction of a file chooser dialog that
uses the running platform's native file chooser dialog. Although
Boxes is Linux-exclusive, we need FileChooserNative in order to
consume the org.freedesktop.portal.FileChooser portal whenever it
is available on the session bus.

This way we bring up an out-of-process file chooser that enables
the user to select a file to be used in Boxes without giving us
access to the filesystem out of the Flatpak sandbox.

These changes introduce some minor regressions such:

* We can no longer embed the file chooser into a Properties page:
For that we perform a little trick of hiding the wizard whenever
the file chooser is visible and vice-versa.

* We can't use file filters or custom filters;