shared-folder-popover: Use GtkFileChooserNative
These changes are necessary in order to make Boxes consume the
Flatpak file chooser portal.

FileChooserNative is an abstraction of a file chooser dialog that
uses the running platform's native file chooser dialog. Although
Boxes is Linux-exclusive, we need FileChooserNative in order to
consume the org.freedesktop.portal.FileChooser portal whenever it
is available on the session bus.

This way we bring up an out-of-process file chooser that enables
the user to select a file to be used in Boxes without giving us
access to the filesystem out of the Flatpak sandbox.

These changes introduced some minor regressions in UX, such as
having a "dialog in a dialog", by having a button that opens the
file chooser instead of a file chooser button as we used to have.
Still, this is compatible with the shared folders mockups available
at https://raw.githubusercontent.com/gnome-design-team/
gnome-mockups/master/boxes/wires/file-sharing.png