wizard-downloads-search: Show recommended OSes as search results
The recommended downloads list should have a precedence over the
Osinfo downloadable filter. So if a vendor wants to enforce some
OS to appear, that doesn't need to satisfy the current filters.

This also fixes the problem of having OSes such as RHEL not
appearing in the search results, since they don't have an explicit
URL for download, but require a custom handling.

Fixes #249