wizard: Don't parse the URIs twice
We are identifying the schemes using two different APIs. GFile's
and libxml. This causes issues with specific protocols being
mapped to others differently.

For instance, an "ssh://" URI is mapped to "sftp://" by the GFile
API. These are two different sources for Boxes. The first is an
SSH connection while the later refers to a secure file transfer
operation.