flatpak: Build QEMU with Virgil support
QEMU can be compiled with additional support for Virgil which is a
solution that allows it to access the OpenGL capabilities of the
host machine's graphics adapter.

This way, the machine emulated by QEMU essentially uses the actual
hardware graphics card for accelerating graphics rendering.

See https://virgil3d.github.io

Fixes #83