installed-media: Drop special handling for gnome-continuos images
It is been a long time since GNOME produces useful continuos images/
composes that can be booted.

Nowadays with Flatpak apps can be tested easily and there's not
much of a demand for a nightly image compose. There are initiatives
to get system components distributed as Flatpaks as well, voiding
the need of GNOME composed images.