list-view: Allow ListView to shrink
We were forcing a constant width for ListView rows in order to limit
the maximium width of the content. This prevented the main Boxes
window from being resized to values smaller than that.

Now we store a max-content-width value in the ListView and whenever
the allocated width for the main window is smaller than that, we
just let the listbox fill the available space.

This is somewhat a responsive behavior for width allocation that
should cover small screen devices or tiling displays.

Hopefully Gtk4 will introduce a max-width CSS property that would
ease the implementation of these use cases. Alternatively, Gtk4
might have a constraint layout machinery that would obsolete this
issue.

Fixes #76