app-window: Don't quit on Ctrl+Q while on Display page
In Gtk the key event handling is top-down in the widget hierarchy.
Therefore the AppWindow object would handle Ctrl+Q before any
widget under it. This is unwanted because most of the time while
on Display mode, users are actually intenting to use the Ctrl+Q
shortcut against the guest, not the host. This would close the
Boxes window and suspend the currently running VMs (very
annoying).

Since we can't guarantee whether the guest properly handled the
key press event as desired, lets just not handle it all in Display
mode/view.

Related #244