vm-importer: Don't auto import InstalledMedias
When the user selects an InstalledMedia (a qcow, vdmk, vdi disk..),
we used to convert it automatically to qcow2 and copy it to a Boxes'
directory under the XDG_CONFIG_DIR.

Now we leave InstalledMedias in their current location delegating to
the user the responsibility to keep the file there (the corresponding
VM wouldn't work if the disk is moved/renamed/deleted.

Possible regressions include disk-format specific limitations, such
as formats not supporting snapshots.

In a following series of patches we will show an "Import" option
for those that want the InstalledMedias to be converted to qcow2
and copied to the Boxes' images location.

These changes are related to #273, because later we would enable
users to "Export" an InstalledMedia (installed VM).