unattended-installer: Pass a better keyboard id to osinfo
We used to pass the same LANG identifier for both language and
keyboard.

These changes lookup into the user system's default input source
and uses it only if it is available in the osinfo-db keyboard
datamap. If not, we search the database for a keyboard that
matches the user's LANG. If none of these two cases work, we
just use the literal LANG string as before.

This should fix #240 and https://bugzilla.redhat.com/1666320