unattended-installer: Pass a better keyboard id to osinfo
We used to pass the same LANG identifier for both language and
keyboard.

These changes lookup on the user's default keyboard layout and
uses it for the express installs. We also make sure that the
current keyboard set in the user's system's is supported by osinfo
by checking against the database.