media-manager: Only express-install supported medias
libosinfo has realised that having installer scripts tied to an OS may
be a mistake and have added those bits to the medias themselves (since
1.4.0).

Trying to work around this very same issue, libosinfo has also added a
new API on OsinfoMedia level that tells whether a media supports an
installer script (since 1.3.0).

Now, with those two new APIs and with the new representation Boxes can
be more accurate on offering express-installations for its users, based
on whether the media itself support the installer script for that
specific profile.

Of course, we're not breaking any kind of compatibility in case we're
running against distros which do not have a newer enough version of
osinfo-db or libosinfo. On those cases, the old (and kind of broken)
behaviour will be kept the same.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>