unattended-installer: Improve need_user_input checks
Although the current check for whether the user input is needed for VM
creation is functional, it's far from being optimal.

More than that, the concept of "need user input" is confusing as it's
related, in the unattended-installer case, to showing to the users the
page where they will select whether to perform a normal or express
installation.

With that in mind, let's do a proper check on the media and OS level to
decide whether the option to provide the express installation should be
displayed or not.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>