unattended-file,archive-writer: Improve initrd-injection support
Some ubuntu initrds are actually multiple cpio archives concatenated,
basic use of cpio -id/libarchive are only able to read a single file
from them, so the approach taken by boxes is not working for them.

virt-install is appending a gzipped single-element cpio archive to the
initrd instead and this commit changes the initrd-injection to do the
same.

Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>